Project renamed to kubectl-gluster
kubectl plugins are supported from 1.13 version,
https://kubernetes.io/docs/tasks/extend-kubectl/kubectl-plugins/
